+++ packages/components-js/tests/vrt/puppeteer/specs/component-scaling.vrt.ts
@@ -1,19 +1,36 @@
-import { getVisualRegressionStatesTester, vrtTest } from '@porsche-design-system/shared/testing';
+import {
+  getVisualRegressionStatesTester,
+  openPopoversAndHighlightSpacer,
+  vrtTest,
+} from '@porsche-design-system/shared/testing';
 
 const components = [
+  'banner',
   'button',
+  'button-pure',
   'checkbox-wrapper',
+  'inline-notification',
   'link',
+  'link-pure',
+  'link-social',
+  'popover',
   'radio-button-wrapper',
   'select-wrapper',
+  'spinner',
+  'tag',
   'tag-dismissible',
   'text-field-wrapper',
+  'textarea-wrapper',
+  'toast-basic',
 ];
 
 it.each(components)('should have no visual regression for scaled component %s', async (component) => {
   expect(
     await vrtTest(getVisualRegressionStatesTester(), `${component}-scaling`, `/#${component}`, {
       scenario: async (page) => {
+        if (component === 'popover') {
+          await openPopoversAndHighlightSpacer(page);
+        }
         const client = await page.target().createCDPSession();
         await client.send('Page.enable');
         await client.send('Page.setFontSizes', {