+++ packages/components-js/tests/vrt/puppeteer/specs/select-wrapper-dropdown.vrt.ts
@@ -10,7 +10,7 @@ it('should have no visual regression for :hover', async () => {
       const page = vrt.getPage();
 
       const head = `<style>
-  p-select-wrapper-dropdown { --p-dropdown-position: 'static'; }
+  p-select-wrapper-dropdown { --p-internal-dropdown-position: 'static'; }
   p-select-wrapper-dropdown:not(:last-child) { margin-bottom: 1rem; }
 </style>`;