+++ packages/components/src/components/banner/banner-styles.ts
@@ -1,58 +1,29 @@
 import type { JssStyle } from 'jss';
+import type { BannerWidth } from './banner-utils';
+import type { KeyframesDirection } from './banner-styles-shared';
 import { getMediaQueryMin, getMediaQueryMinMax } from '@porsche-design-system/utilities-v2';
-import { buildSlottedStyles, getCss } from '../../utils';
-import { addImportantToRule, getBaseSlottedStyles, pxToRemWithUnit } from '../../styles';
+import { getCss } from '../../utils';
 import { BANNER_Z_INDEX } from '../../constants';
+import { getContentWrapperStyle } from '../content-wrapper/content-wrapper-styles-shared';
+import {
+  getAnimationIn,
+  getAnimationOut,
+  getBoxShadow,
+  getKeyframes,
+  getKeyframesMobile,
+} from './banner-styles-shared';
+import { addImportantToEachRule } from '../../styles';
 
 const bannerPositionTypeVar = '--p-banner-position-type';
 const bannerPositionTopVar = '--p-banner-position-top';
 const bannerPositionBottomVar = '--p-banner-position-bottom';
-const bannerZIndexVar = '--p-banner-z-index';
-const bannerAnimationDurationVar = '--p-animation-duration__banner';
+const bannerZIndexVar = '--p-internal-banner-z-index';
+const bannerAnimationDurationVar = '--p-animation-duration';
 
-const easeInQuad = 'cubic-bezier(0.45,0,0.55,1)';
-const easeOutQuad = 'cubic-bezier(0.5,1,0.89,1)';
-export const ANIMATION_DURATION = 600;
+const bannerOffset = '56px';
 
-const mediaQueryS = getMediaQueryMin('s');
 const mediaQueryBase = getMediaQueryMinMax('base', 's');
-
-export const getBoxShadow = (): JssStyle => ({
-  boxShadow:
-    `0 ${pxToRemWithUnit(2)} ${pxToRemWithUnit(4)} 0 rgba(0,0,0,0.05),` +
-    `0 ${pxToRemWithUnit(15)} ${pxToRemWithUnit(20)} 0 rgba(0,0,0,0.2)`,
-});
-
-export const getAnimationIn = (keyframesName: string, durationVar?: string): JssStyle => {
-  const duration = durationVar ? `var(${durationVar},${ANIMATION_DURATION}ms)` : `${ANIMATION_DURATION}ms`;
-  return {
-    animation: `${duration} $${keyframesName} ${easeInQuad} forwards`,
-  };
-};
-
-export const getAnimationOut = (keyframesName: string): JssStyle => ({
-  animation: addImportantToRule(`${ANIMATION_DURATION}ms $${keyframesName} ${easeOutQuad} forwards`),
-});
-
-export type KeyframesDirection = 'in' | 'out';
-const getKeyframes = (direction: KeyframesDirection, outsideStyle: JssStyle): JssStyle => {
-  const insideStyle: JssStyle = { opacity: 1, transform: 'translate3d(0,0,0)' };
-  return direction === 'in'
-    ? {
-        from: outsideStyle,
-        to: insideStyle,
-      }
-    : {
-        from: insideStyle,
-        to: outsideStyle,
-      };
-};
-
-export const getKeyframesMobile = (direction: KeyframesDirection, bottomVar: string): JssStyle =>
-  getKeyframes(direction, {
-    opacity: 0,
-    transform: `translate3d(0,calc(var(${bottomVar}) + 100%),0)`, // space before and after "+" is crucial
-  });
+const mediaQueryS = getMediaQueryMin('s');
 
 const getKeyframesDesktop = (direction: KeyframesDirection, topVar: string): JssStyle =>
   getKeyframes(direction, {
@@ -60,13 +31,13 @@ const getKeyframesDesktop = (direction: KeyframesDirection, topVar: string): Jss
     transform: `translate3d(0,calc(-100% - var(${topVar})),0)`, // space before and after "-" is crucial
   });
 
-export const getComponentCss = (): string => {
+export const getComponentCss = (width: BannerWidth): string => {
   return getCss({
     '@global': {
       ':host': {
         // TODO: Why is nothing set as important here?
-        [bannerPositionTopVar]: pxToRemWithUnit(56),
-        [bannerPositionBottomVar]: pxToRemWithUnit(56),
+        [bannerPositionTopVar]: bannerOffset,
+        [bannerPositionBottomVar]: bannerOffset,
         display: 'block',
         position: `var(${bannerPositionTypeVar},fixed)`,
         zIndex: `var(${bannerZIndexVar},${BANNER_Z_INDEX})`,
@@ -74,6 +45,7 @@ export const getComponentCss = (): string => {
         left: 0,
         right: 0,
         willChange: 'opacity,transform',
+        ...addImportantToEachRule(getContentWrapperStyle(width)),
         [mediaQueryBase]: {
           bottom: `var(${bannerPositionBottomVar})`,
         },
@@ -96,8 +68,4 @@ export const getComponentCss = (): string => {
     },
     root: getBoxShadow(),
   });
-};
-
-export const getSlottedCss = (host: HTMLElement): string => {
-  return getCss(buildSlottedStyles(host, getBaseSlottedStyles()));
 };