+++ packages/components/src/components/select-wrapper/select-wrapper/select-wrapper.tsx
@@ -170,7 +170,7 @@ export class SelectWrapper {
           )}
         </div>
         {hasMessage(this.host, this.message, this.state) && (
-          <StateMessage state={this.state} message={this.message} host={this.host} />
+          <StateMessage state={this.state} message={this.message} theme={this.theme} host={this.host} />
         )}
       </Host>
     );