+++ packages/components/src/components/toast/toast/toast-manager.ts
@@ -1,13 +1,12 @@
 import { forceUpdate } from '@stencil/core';
 import type { ToastState } from './toast-utils';
-import { ANIMATION_DURATION } from '../../banner/banner-styles';
+import { ANIMATION_DURATION } from '../../banner/banner-styles-shared';
 
 const TOAST_DEFAULT_TIMEOUT = 6000;
 
 // css variable names for overriding behaviour in tests
-const TOAST_CSS_SKIP_TIMEOUT_VAR = '--p-override-toast-skip-timeout';
-const TOAST_CSS_TIMEOUT_OVERRIDE_VAR = '--p-override-toast-timeout';
-export const TOAST_ANIMATION_DURATION_VAR = '--p-override-toast-animation-duration';
+const TEMPORARY_TOAST_SKIP_TIMEOUT = '--p-temporary-toast-skip-timeout';
+const TEMPORARY_TOAST_TIMEOUT = '--p-temporary-toast-timeout';
 
 export type ToastMessage = {
   text: string;
@@ -43,7 +42,7 @@ export class ToastManagerClass {
       throw new Error('Empty text provided to addMessage.');
     }
 
-    const msg: ToastMessage = { state: 'neutral', ...message }; // neutral is our default state
+    const msg: ToastMessage = { state: 'info', ...message }; // info is our default state
 
     const { length } = this.messages;
     this.messages.push(msg);
@@ -59,11 +58,10 @@ export class ToastManagerClass {
     this.onDismissCallback();
     setTimeout(
       () => forceUpdate(this.toastEl),
-      // respect --p-override-toast-animation-duration css variable to override timeout during e2e and vrt tests
+      // respect --p-temporary-toast-timeout css variable to override timeout during e2e and vrt tests
       ROLLUP_REPLACE_IS_STAGING === 'production' || process.env.NODE_ENV === 'test'
         ? ANIMATION_DURATION
-        : parseInt(getComputedStyle(this.toastEl).getPropertyValue(TOAST_ANIMATION_DURATION_VAR), 10) ||
-            ANIMATION_DURATION
+        : parseInt(getComputedStyle(this.toastEl).getPropertyValue('--p-animation-duration'), 10) || ANIMATION_DURATION
     );
   };
 
@@ -77,12 +75,12 @@ export class ToastManagerClass {
       if (ROLLUP_REPLACE_IS_STAGING === 'production' || process.env.NODE_ENV === 'test') {
         this.timeout = setTimeout(this.dismissToastItem, TOAST_DEFAULT_TIMEOUT);
       } else {
-        // skip setting timeout if --p-override-toast-skip-timeout css variable is set in dev build
-        if (getComputedStyle(this.toastEl).getPropertyValue(TOAST_CSS_SKIP_TIMEOUT_VAR)?.trim() !== 'true') {
+        // skip setting timeout if --p-temporary-toast-skip-timeout css variable is set in dev build
+        if (getComputedStyle(this.toastEl).getPropertyValue(TEMPORARY_TOAST_SKIP_TIMEOUT)?.trim() !== 'true') {
           this.timeout = setTimeout(
             this.dismissToastItem,
-            // override timeout if --p-override-toast-timeout css variable is set
-            parseInt(getComputedStyle(this.toastEl).getPropertyValue(TOAST_CSS_TIMEOUT_OVERRIDE_VAR), 10) ||
+            // override timeout if --p-temporary-toast-timeout css variable is set
+            parseInt(getComputedStyle(this.toastEl).getPropertyValue(TEMPORARY_TOAST_TIMEOUT), 10) ||
               TOAST_DEFAULT_TIMEOUT
           );
         }