+++ packages/components/src/components/toast/toast/toast-styles.ts
@@ -1,12 +1,11 @@
-import { gridSafeZone, getMediaQueryMin } from '@porsche-design-system/utilities-v2';
+import { getMediaQueryMin, gridSafeZone } from '@porsche-design-system/utilities-v2';
 import { getCss } from '../../../utils';
-import { addImportantToEachRule, pxToRemWithUnit } from '../../../styles';
+import { addImportantToEachRule } from '../../../styles';
 import { TOAST_Z_INDEX } from '../../../constants';
-import { getAnimationIn, getAnimationOut, getKeyframesMobile } from '../../banner/banner-styles';
-import { TOAST_ANIMATION_DURATION_VAR } from './toast-manager';
+import { getAnimationIn, getAnimationOut, getKeyframesMobile } from '../../banner/banner-styles-shared';
 
 const toastPositionBottomVarPublic = '--p-toast-position-bottom'; // CSS custom property exposed as public interface
-const toastPositionBottomVarInternal = `${toastPositionBottomVarPublic}-internal`;
+const toastPositionBottomVarInternal = '--p-internal-toast-position-bottom';
 export const toastCloseClassName = 'close';
 
 export const getComponentCss = (): string => {
@@ -16,16 +15,16 @@ export const getComponentCss = (): string => {
         position: 'fixed',
         left: gridSafeZone,
         right: gridSafeZone,
-        // Needs a not overwritable internal css variable to cover default position depending on viewport size and to handle animation properly.
+        // Needs a not overridable internal css variable to cover default position depending on viewport size and to handle animation properly.
         // In addition, a public css variable can be passed to overwrite the default position.
-        [toastPositionBottomVarInternal]: `var(${toastPositionBottomVarPublic}, ${pxToRemWithUnit(56)})`,
+        [toastPositionBottomVarInternal]: `var(${toastPositionBottomVarPublic}, 56px)`,
         bottom: `var(${toastPositionBottomVarInternal})`,
         maxWidth: '42rem',
         zIndex: TOAST_Z_INDEX,
         [getMediaQueryMin('s')]: {
-          left: pxToRemWithUnit(64),
+          left: '64px',
           right: 'auto',
-          [toastPositionBottomVarInternal]: `var(${toastPositionBottomVarPublic}, ${pxToRemWithUnit(64)})`,
+          [toastPositionBottomVarInternal]: `var(${toastPositionBottomVarPublic}, 64px)`,
           bottom: `var(${toastPositionBottomVarInternal})`,
         },
       }),
@@ -34,7 +33,7 @@ export const getComponentCss = (): string => {
     },
     hydrated: getAnimationIn(
       'in',
-      ROLLUP_REPLACE_IS_STAGING !== 'production' && process.env.NODE_ENV !== 'test' && TOAST_ANIMATION_DURATION_VAR
+      ROLLUP_REPLACE_IS_STAGING !== 'production' && process.env.NODE_ENV !== 'test' && '--p-animation-duration'
     ),
     [toastCloseClassName]: getAnimationOut('out'),
   });