+++ packages/components/src/styles/form-styles.ts
@@ -1,17 +1,19 @@
 import type { JssStyle, Styles } from 'jss';
 import type { BreakpointCustomizable, Theme } from '../types';
-import { buildResponsiveStyles, isVisibleFormState } from '../utils';
-import { addImportantToRule, getFormTextHiddenJssStyle, getThemedColors, getTransition, pxToRemWithUnit } from './';
+import { buildResponsiveStyles } from '../utils';
+import { addImportantToRule, getFormTextHiddenJssStyle, getThemedColors, getTransition } from './';
 import {
   borderRadiusSmall,
-  textSmallStyle,
+  borderWidthBase,
   fontSizeTextXSmall,
   spacingStaticXSmall,
+  textSmallStyle,
 } from '@porsche-design-system/utilities-v2';
 import { getThemedFormStateColors } from './form-state-color-styles';
 import { hoverMediaQuery } from './hover-media-query';
 import type { FormState } from '../utils/form/form-state';
 
+// TODO: should be removed if possible?
 export const INPUT_HEIGHT = 54;
 
 export type ChildSelector = 'input' | 'select' | 'textarea';
@@ -22,42 +24,36 @@ export const getBaseChildStyles = (
   theme: Theme,
   additionalDefaultJssStyle?: JssStyle
 ): Styles => {
-  const { primaryColor, contrastLowColor, contrastMediumColor, disabledColor, focusColor } = getThemedColors(theme);
+  const { primaryColor, contrastLowColor, contrastMediumColor, disabledColor } = getThemedColors(theme);
   const { formStateColor, formStateHoverColor } = getThemedFormStateColors(theme, state);
-  const hasVisibleState = isVisibleFormState(state);
 
   return {
     [`::slotted(${child})`]: {
       display: 'block',
-      position: 'relative',
       width: '100%',
-      ...(child !== 'textarea' && { height: pxToRemWithUnit(INPUT_HEIGHT) }),
-      outline: '2px solid transparent',
-      outlineOffset: '2px',
+      height: 'auto',
+      margin: 0,
+      outline: 0,
       WebkitAppearance: 'none', // iOS safari
       appearance: 'none',
       boxSizing: 'border-box',
-      border: `2px solid ${hasVisibleState ? formStateColor : contrastMediumColor}`,
+      border: `${borderWidthBase} solid ${formStateColor || contrastMediumColor}`,
       borderRadius: borderRadiusSmall,
       background: 'transparent',
-      font: textSmallStyle.font,
+      font: textSmallStyle.font.replace('ex', 'ex + 10px'), // a minimum line-height is needed for input, otherwise value is scrollable in Chrome
       textIndent: 0,
       color: primaryColor,
-      transition: getTransition('border-color'),
+      transition: ['color', 'border-color', 'background-color'].map(getTransition).join(), // for smooth transitions between e.g. disabled states
       ...additionalDefaultJssStyle,
     },
     ...(hoverMediaQuery({
       // with the media query the selector has higher priority and overrides disabled styles
-      [`::slotted(${child}:not(:disabled):not([readonly]):hover)`]: {
+      [`::slotted(${child}:not(:disabled):not(:focus):not([readonly]):hover)`]: {
         borderColor: formStateHoverColor || primaryColor,
       },
     }) as Styles),
     [`::slotted(${child}:focus)`]: {
-      borderColor: formStateColor || primaryColor,
-      outlineColor: focusColor,
-    },
-    [`::slotted(${child}:focus:not(:focus-visible))`]: {
-      outlineColor: 'transparent',
+      borderColor: primaryColor,
     },
     [`::slotted(${child}:disabled)`]: {
       cursor: 'not-allowed',
@@ -84,20 +80,21 @@ export const getLabelStyles = (
 ): Styles => {
   const { primaryColor, disabledColor, contrastHighColor } = getThemedColors(theme);
   const { formStateHoverColor } = getThemedFormStateColors(theme, state);
-  const hasVisibleState = isVisibleFormState(state);
 
   const counterOrUnitOrIconStylesKey = counterOrUnitOrIconStyles && Object.keys(counterOrUnitOrIconStyles)[0];
 
   return {
     label: {
-      display: 'block',
+      display: 'flex',
+      flexDirection: 'column',
+      gap: spacingStaticXSmall,
       position: 'relative', // for unit and counter
       '&__text': {
         display: 'block',
         ...buildResponsiveStyles(hideLabel, getFormTextHiddenJssStyle),
         ...textSmallStyle,
         color: isDisabled ? disabledColor : primaryColor,
-        // transition: getTransition('color'),
+        transition: getTransition('color'), // for smooth transitions between e.g. disabled states
         '&+&': {
           marginTop: `-${spacingStaticXSmall}`,
           fontSize: fontSizeTextXSmall,
@@ -107,9 +104,9 @@ export const getLabelStyles = (
         },
         ...hoverMediaQuery({
           '&:hover': {
-            [`&~::slotted(${child}:not(:disabled):not([readonly]))` +
-            (hasVisibleState ? `,::slotted(${child}:not(:disabled):not([readonly]):hover)` : '')]: {
-              borderColor: addImportantToRule(hasVisibleState ? formStateHoverColor : primaryColor),
+            [`&~::slotted(${child}:not(:disabled):not(:focus):not([readonly]))` +
+            (formStateHoverColor ? `,::slotted(${child}:not(:disabled):not(:focus):not([readonly]):hover)` : '')]: {
+              borderColor: addImportantToRule(formStateHoverColor || primaryColor),
             },
           },
         }),